Bijimio Experimental Fork V6.7.7 | Adjusted and added some missing attributes
