+++ src/main/java/net/ltxprogrammer/changed/entity/beast/AbstractSnowLeopard.java
@@ -5,7 +5,9 @@ import net.ltxprogrammer.changed.util.Color3;
 import net.minecraft.world.damagesource.DamageSource;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.ai.attributes.AttributeMap;
+import net.minecraft.world.entity.ai.attributes.Attributes;
 import net.minecraft.world.level.Level;
+import net.minecraftforge.common.ForgeMod;
 
 public abstract class AbstractSnowLeopard extends ChangedEntity implements GenderedEntity, PowderSnowWalkable {
     public AbstractSnowLeopard(EntityType<? extends AbstractSnowLeopard> p_19870_, Level p_19871_) {
@@ -38,6 +40,9 @@ public abstract class AbstractSnowLeopard extends ChangedEntity implements Gende
     @Override
     protected void setAttributes(AttributeMap attributes) {
         super.setAttributes(attributes);
-        AttributePresets.catLike(attributes);
+        attributes.getInstance(Attributes.MOVEMENT_SPEED).setBaseValue(0.12f);
+        attributes.getInstance(ForgeMod.SWIM_SPEED.get()).setBaseValue(0.9);
+        attributes.getInstance(Attributes.MAX_HEALTH).setBaseValue(22.0);
     }
+
 }
