+++ src/main/java/net/ltxprogrammer/changed/entity/beast/DarkLatexDragon.java
@@ -3,6 +3,7 @@ package net.ltxprogrammer.changed.entity.beast;
 import net.ltxprogrammer.changed.entity.*;
 import net.ltxprogrammer.changed.util.Color3;
 import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.ai.attributes.AttributeMap;
 import net.minecraft.world.level.Level;
 import org.jetbrains.annotations.Nullable;
 
@@ -15,6 +16,12 @@ public class DarkLatexDragon extends ChangedEntity implements DarkLatexEntity, P
 
     public boolean isMaskless() {
         return true;
+    }
+
+    @Override
+    protected void setAttributes(AttributeMap attributes) {
+        super.setAttributes(attributes);
+        AttributePresets.dragonLike(attributes);
     }
 
     @Override
