+++ src/main/java/net/ltxprogrammer/changed/entity/beast/LatexBeifeng.java
@@ -19,6 +19,13 @@ public class LatexBeifeng extends AbstractLatexWolf implements DarkLatexEntity, 
     }
 
     @Override
+    protected void setAttributes(AttributeMap attributes) {
+        super.setAttributes(attributes);
+        attributes.getInstance(Attributes.MOVEMENT_SPEED).setBaseValue(0.105);
+        attributes.getInstance(ForgeMod.SWIM_SPEED.get()).setBaseValue(1.0);
+    }
+
+    @Override
     public Color3 getDripColor() {
         return level.random.nextInt(10) > 3 ? Color3.BLUE : Color3.WHITE;
     }
@@ -39,12 +46,5 @@ public class LatexBeifeng extends AbstractLatexWolf implements DarkLatexEntity, 
 
     public Color3 getTransfurColor(TransfurCause cause) {
         return Color3.getColor("#51659d");
-    }
-
-    @Override
-    protected void setAttributes(AttributeMap attributes) {
-        super.setAttributes(attributes);
-        attributes.getInstance(Attributes.MOVEMENT_SPEED).setBaseValue(0.105);
-        attributes.getInstance(ForgeMod.SWIM_SPEED.get()).setBaseValue(1.0);
     }
 }
