+++ src/main/java/net/ltxprogrammer/changed/entity/beast/LatexBlueDragon.java
@@ -17,6 +17,13 @@ public class LatexBlueDragon extends ChangedEntity implements PowderSnowWalkable
     }
 
     @Override
+    protected void setAttributes(AttributeMap attributes) {
+        super.setAttributes(attributes);
+        attributes.getInstance(Attributes.MOVEMENT_SPEED).setBaseValue(0.105);
+        attributes.getInstance(ForgeMod.SWIM_SPEED.get()).setBaseValue(1.0);
+    }
+
+    @Override
     public int getTicksRequiredToFreeze() { return 480; }
 
     @Override
@@ -45,12 +52,5 @@ public class LatexBlueDragon extends ChangedEntity implements PowderSnowWalkable
 
     public Color3 getTransfurColor(TransfurCause cause) {
         return Color3.getColor("#eef9ff");
-    }
-
-    @Override
-    protected void setAttributes(AttributeMap attributes) {
-        super.setAttributes(attributes);
-        attributes.getInstance(Attributes.MOVEMENT_SPEED).setBaseValue(0.105);
-        attributes.getInstance(ForgeMod.SWIM_SPEED.get()).setBaseValue(0.98);
     }
 }
