+++ src/main/java/net/ltxprogrammer/changed/entity/beast/LatexKeonWolf.java
@@ -1,10 +1,12 @@
 package net.ltxprogrammer.changed.entity.beast;
 
+import net.ltxprogrammer.changed.entity.AttributePresets;
 import net.ltxprogrammer.changed.entity.HairStyle;
 import net.ltxprogrammer.changed.entity.PatronOC;
 import net.ltxprogrammer.changed.entity.TransfurCause;
 import net.ltxprogrammer.changed.util.Color3;
 import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.ai.attributes.AttributeMap;
 import net.minecraft.world.level.Level;
 import org.jetbrains.annotations.Nullable;
 
@@ -14,6 +16,13 @@ public class LatexKeonWolf extends AbstractLatexWolf implements PatronOC {
     public LatexKeonWolf(EntityType<? extends LatexKeonWolf> p_19870_, Level p_19871_) {
         super(p_19870_, p_19871_);
     }
+
+    @Override
+    protected void setAttributes(AttributeMap attributes) {
+        super.setAttributes(attributes);
+        AttributePresets.wolfLike(attributes);
+    }
+
     @Override
     public Color3 getDripColor() {
         return Color3.SILVER;
