+++ src/main/java/net/ltxprogrammer/changed/entity/beast/LatexMantaRayFemale.java
@@ -21,6 +21,14 @@ public class LatexMantaRayFemale extends AbstractLatexMantaRay {
     }
 
     @Override
+    protected void setAttributes(AttributeMap attributes) {
+        super.setAttributes(attributes);
+        attributes.getInstance(Attributes.MOVEMENT_SPEED).setBaseValue(0.034);
+        attributes.getInstance(ForgeMod.SWIM_SPEED.get()).setBaseValue(1.9);
+        attributes.getInstance(Attributes.MAX_HEALTH).setBaseValue(28);
+    }
+
+    @Override
     public Gender getGender() {
         return Gender.FEMALE;
     }
@@ -44,13 +52,5 @@ public class LatexMantaRayFemale extends AbstractLatexMantaRay {
 
     public Color3 getTransfurColor(TransfurCause cause) {
         return Color3.getColor("#6f7696");
-    }
-
-    @Override
-    protected void setAttributes(AttributeMap attributes) {
-        super.setAttributes(attributes);
-        attributes.getInstance(Attributes.MOVEMENT_SPEED).setBaseValue(0.026);
-        attributes.getInstance(ForgeMod.SWIM_SPEED.get()).setBaseValue(1.9);
-        attributes.getInstance(Attributes.MAX_HEALTH).setBaseValue(28);
     }
 }
