+++ src/main/java/net/ltxprogrammer/changed/entity/beast/LatexMimicPlant.java
@@ -19,6 +19,13 @@ public class LatexMimicPlant extends LatexLeaf {
     }
 
     @Override
+    protected void setAttributes(AttributeMap attributes) {
+        super.setAttributes(attributes);
+        attributes.getInstance(Attributes.MOVEMENT_SPEED).setBaseValue(0.105);
+        attributes.getInstance(ForgeMod.SWIM_SPEED.get()).setBaseValue(0.95);
+    }
+
+    @Override
     public TransfurMode getTransfurMode() {
         return TransfurMode.ABSORPTION;
     }
@@ -44,12 +51,5 @@ public class LatexMimicPlant extends LatexLeaf {
 
     public Color3 getTransfurColor(TransfurCause cause) {
         return Color3.getColor("#497469");
-    }
-
-    @Override
-    protected void setAttributes(AttributeMap attributes) {
-        super.setAttributes(attributes);
-        attributes.getInstance(Attributes.MOVEMENT_SPEED).setBaseValue(0.105);
-        attributes.getInstance(ForgeMod.SWIM_SPEED.get()).setBaseValue(0.95);
     }
 }
