+++ src/main/java/net/ltxprogrammer/changed/entity/beast/LatexMoth.java
@@ -17,6 +17,13 @@ public class LatexMoth extends ChangedEntity {
     }
 
     @Override
+    protected void setAttributes(AttributeMap attributes) {
+        super.setAttributes(attributes);
+        attributes.getInstance(Attributes.MOVEMENT_SPEED).setBaseValue(0.105);
+        attributes.getInstance(ForgeMod.SWIM_SPEED.get()).setBaseValue(0.75);
+    }
+
+    @Override
     public int getTicksRequiredToFreeze() { return 240; }
 
     @Override
@@ -49,12 +56,5 @@ public class LatexMoth extends ChangedEntity {
 
     public Color3 getTransfurColor(TransfurCause cause) {
         return Color3.getColor("#d8bc99");
-    }
-
-    @Override
-    protected void setAttributes(AttributeMap attributes) {
-        super.setAttributes(attributes);
-        attributes.getInstance(Attributes.MOVEMENT_SPEED).setBaseValue(0.105);
-        attributes.getInstance(ForgeMod.SWIM_SPEED.get()).setBaseValue(0.75);
     }
 }
