+++ src/main/java/net/ltxprogrammer/changed/entity/beast/LatexMutantBloodcellWolf.java
@@ -13,14 +13,14 @@ public class LatexMutantBloodcellWolf extends WhiteLatexEntity {
     }
 
     @Override
-    public TransfurMode getTransfurMode() {
-        return TransfurMode.ABSORPTION;
+    protected void setAttributes(AttributeMap attributes) {
+        super.setAttributes(attributes);
+        attributes.getInstance(Attributes.MOVEMENT_SPEED).setBaseValue(0.1);
+        attributes.getInstance(ForgeMod.SWIM_SPEED.get()).setBaseValue(0.9);
     }
 
     @Override
-    protected void setAttributes(AttributeMap attributes) {
-        super.setAttributes(attributes);
-        attributes.getInstance(Attributes.MOVEMENT_SPEED).setBaseValue(0.105);
-        attributes.getInstance(ForgeMod.SWIM_SPEED.get()).setBaseValue(0.9);
+    public TransfurMode getTransfurMode() {
+        return TransfurMode.ABSORPTION;
     }
 }
