+++ src/main/java/net/ltxprogrammer/changed/entity/beast/LatexPinkDeer.java
@@ -8,7 +8,9 @@ import net.ltxprogrammer.changed.entity.TransfurMode;
 import net.ltxprogrammer.changed.util.Color3;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.ai.attributes.AttributeMap;
+import net.minecraft.world.entity.ai.attributes.Attributes;
 import net.minecraft.world.level.Level;
+import net.minecraftforge.common.ForgeMod;
 
 
 public class LatexPinkDeer extends LatexPinkWyvern {
@@ -16,6 +18,12 @@ public class LatexPinkDeer extends LatexPinkWyvern {
         super(p_19870_, p_19871_);
     }
 
+    @Override
+    protected void setAttributes(AttributeMap attributes) {
+        super.setAttributes(attributes);
+        attributes.getInstance(Attributes.MOVEMENT_SPEED).setBaseValue(0.1075);
+        attributes.getInstance(ForgeMod.SWIM_SPEED.get()).setBaseValue(0.95);
+    }
 
     @Override
     public Color3 getDripColor() {
@@ -40,11 +48,5 @@ public class LatexPinkDeer extends LatexPinkWyvern {
             return Color3.fromInt(0xd8bc99);
         else
             return Color3.fromInt(0xf7aebe);
-    }
-
-    @Override
-    protected void setAttributes(AttributeMap attributes) {
-        super.setAttributes(attributes);
-        AttributePresets.wolfLike(attributes);
     }
 }
