+++ src/main/java/net/ltxprogrammer/changed/entity/beast/LatexPinkWyvern.java
@@ -18,6 +18,13 @@ public class LatexPinkWyvern extends ChangedEntity implements PowderSnowWalkable
     }
 
     @Override
+    protected void setAttributes(AttributeMap attributes) {
+        super.setAttributes(attributes);
+        attributes.getInstance(Attributes.MOVEMENT_SPEED).setBaseValue(0.105);
+        attributes.getInstance(ForgeMod.SWIM_SPEED.get()).setBaseValue(0.95);
+    }
+
+    @Override
     public Color3 getDripColor() {
         return Color3.getColor(this.random.nextInt(4) < 3 ? "#f7aebe" : "#ffffff");
     }
@@ -48,13 +55,6 @@ public class LatexPinkWyvern extends ChangedEntity implements PowderSnowWalkable
 
     public Color3 getTransfurColor(TransfurCause cause) {
         return Color3.getColor("#f7aebe");
-    }
-
-    @Override
-    protected void setAttributes(AttributeMap attributes) {
-        super.setAttributes(attributes);
-        attributes.getInstance(Attributes.MOVEMENT_SPEED).setBaseValue(0.105);
-        attributes.getInstance(ForgeMod.SWIM_SPEED.get()).setBaseValue(0.95);
     }
 }
