+++ src/main/java/net/ltxprogrammer/changed/entity/beast/LatexSiren.java
@@ -27,6 +27,14 @@ public class LatexSiren extends AbstractAquaticGenderedEntity {
     }
 
     @Override
+    protected void setAttributes(AttributeMap attributes) {
+        super.setAttributes(attributes);
+        attributes.getInstance(Attributes.MOVEMENT_SPEED).setBaseValue(0.034);
+        attributes.getInstance(ForgeMod.SWIM_SPEED.get()).setBaseValue(1.9);
+        attributes.getInstance(Attributes.MAX_HEALTH).setBaseValue(28);
+    }
+
+    @Override
     public Gender getGender() {
         return Gender.FEMALE;
     }
@@ -64,13 +72,5 @@ public class LatexSiren extends AbstractAquaticGenderedEntity {
 
     public Color3 getTransfurColor(TransfurCause cause) {
         return Color3.getColor("#969696");
-    }
-
-    @Override
-    protected void setAttributes(AttributeMap attributes) {
-        super.setAttributes(attributes);
-        attributes.getInstance(Attributes.MOVEMENT_SPEED).setBaseValue(0.026);
-        attributes.getInstance(ForgeMod.SWIM_SPEED.get()).setBaseValue(1.9);
-        attributes.getInstance(Attributes.MAX_HEALTH).setBaseValue(28);
     }
 }
