+++ src/main/java/net/ltxprogrammer/changed/entity/beast/LatexSnake.java
@@ -20,6 +20,14 @@ public class LatexSnake extends ChangedEntity {
     }
 
     @Override
+    protected void setAttributes(AttributeMap attributes) {
+        super.setAttributes(attributes);
+        attributes.getInstance(Attributes.MOVEMENT_SPEED).setBaseValue(0.1);
+        attributes.getInstance(ForgeMod.SWIM_SPEED.get()).setBaseValue(0.95);
+        attributes.getInstance(Attributes.MAX_HEALTH).setBaseValue(26);
+    }
+
+    @Override
     public LatexType getLatexType() {
         return LatexType.NEUTRAL;
     }
@@ -51,13 +59,5 @@ public class LatexSnake extends ChangedEntity {
     @Override
     public boolean isMovingSlowly() {
         return this.isCrouching();
-    }
-
-    @Override
-    protected void setAttributes(AttributeMap attributes) {
-        super.setAttributes(attributes);
-        attributes.getInstance(Attributes.MOVEMENT_SPEED).setBaseValue(0.1);
-        attributes.getInstance(ForgeMod.SWIM_SPEED.get()).setBaseValue(0.95);
-        attributes.getInstance(Attributes.MAX_HEALTH).setBaseValue(26);
     }
 }
