+++ src/main/java/net/ltxprogrammer/changed/entity/beast/LatexSnowLeopardMale.java
@@ -17,11 +17,6 @@ public class LatexSnowLeopardMale extends AbstractSnowLeopard {
         super(p_19870_, p_19871_);
     }
 
-    protected void setAttributes(AttributeMap attributes) {
-        super.setAttributes(attributes);
-        attributes.getInstance(Attributes.MOVEMENT_SPEED).setBaseValue(1.2f);
-    }
-
     @Override
     public HairStyle getDefaultHairStyle() {
         return HairStyle.BALD.get();
