+++ src/main/java/net/ltxprogrammer/changed/entity/beast/LatexTigerShark.java
@@ -23,6 +23,14 @@ public class LatexTigerShark extends AbstractAquaticEntity {
         summonSharks = registerAbility(ability -> this.wantToSummon(), new SimpleAbilityInstance(ChangedAbilities.SUMMON_SHARKS.get(), IAbstractChangedEntity.forEntity(this)));
     }
 
+    @Override
+    protected void setAttributes(AttributeMap attributes) {
+        super.setAttributes(attributes);
+        attributes.getInstance(Attributes.MOVEMENT_SPEED).setBaseValue(0.0925);
+        attributes.getInstance(ForgeMod.SWIM_SPEED.get()).setBaseValue(1.15);
+        attributes.getInstance(Attributes.MAX_HEALTH).setBaseValue(28.0);
+    }
+
     public boolean wantToSummon() {
         return getTarget() != null;
     }
@@ -43,13 +51,5 @@ public class LatexTigerShark extends AbstractAquaticEntity {
 
     public Color3 getTransfurColor(TransfurCause cause) {
         return Color3.getColor("#969696");
-    }
-
-    @Override
-    protected void setAttributes(AttributeMap attributes) {
-        super.setAttributes(attributes);
-        attributes.getInstance(Attributes.MOVEMENT_SPEED).setBaseValue(0.0925);
-        attributes.getInstance(ForgeMod.SWIM_SPEED.get()).setBaseValue(1.2);
-        attributes.getInstance(Attributes.MAX_HEALTH).setBaseValue(28.0);
     }
 }
