+++ src/main/java/net/ltxprogrammer/changed/entity/beast/WhiteLatexCentaur.java
@@ -23,6 +23,14 @@ public class WhiteLatexCentaur extends WhiteLatexKnight implements LatexTaur<Whi
     }
 
     @Override
+    protected void setAttributes(AttributeMap attributes) {
+        super.setAttributes(attributes);
+        attributes.getInstance(Attributes.MOVEMENT_SPEED).setBaseValue(0.12);
+        attributes.getInstance(ForgeMod.SWIM_SPEED.get()).setBaseValue(0.9);
+        attributes.getInstance(Attributes.MAX_HEALTH).setBaseValue(30);
+    }
+
+    @Override
     public LatexType getLatexType() {
         return LatexType.NEUTRAL;
     }
@@ -86,13 +94,5 @@ public class WhiteLatexCentaur extends WhiteLatexKnight implements LatexTaur<Whi
     @Override
     public boolean isAllowedToSwim() {
         return true;
-    }
-
-    @Override
-    protected void setAttributes(AttributeMap attributes) {
-        super.setAttributes(attributes);
-        attributes.getInstance(Attributes.MOVEMENT_SPEED).setBaseValue(0.12);
-        attributes.getInstance(ForgeMod.SWIM_SPEED.get()).setBaseValue(0.9);
-        attributes.getInstance(Attributes.MAX_HEALTH).setBaseValue(30);
     }
 }
