+++ src/main/java/net/ltxprogrammer/changed/entity/beast/WhiteLatexKnightFusion.java
@@ -10,6 +10,7 @@ import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.ai.attributes.AttributeMap;
 import net.minecraft.world.entity.ai.attributes.Attributes;
 import net.minecraft.world.level.Level;
+import net.minecraftforge.common.ForgeMod;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.List;
@@ -17,6 +18,14 @@ import java.util.List;
 public class WhiteLatexKnightFusion extends WhiteLatexKnight {
     public WhiteLatexKnightFusion(EntityType<? extends WhiteLatexKnightFusion> p_19870_, Level p_19871_) {
         super(p_19870_, p_19871_);
+    }
+
+    @Override
+    protected void setAttributes(AttributeMap attributes) {
+        super.setAttributes(attributes);
+        attributes.getInstance(Attributes.MAX_HEALTH).setBaseValue(28.0);
+        attributes.getInstance(Attributes.MOVEMENT_SPEED).setBaseValue(0.1075f);
+        attributes.getInstance(ForgeMod.SWIM_SPEED.get()).setBaseValue(0.95);
     }
 
     @Override
@@ -51,11 +60,5 @@ public class WhiteLatexKnightFusion extends WhiteLatexKnight {
     @Override
     public TransfurVariant<?> getTransfurVariant() {
         return ChangedTransfurVariants.WHITE_LATEX_KNIGHT_FUSION.get();
-    }
-
-    @Override
-    protected void setAttributes(AttributeMap attributes) {
-        super.setAttributes(attributes);
-        attributes.getInstance(Attributes.MAX_HEALTH).setBaseValue(28.0);
     }
 }
